+++ src/events/index.ts
@@ -1,7 +1,6 @@
 import { Event } from '../lib';
 
 import ready from './ready';
-import scamLogs from './logging/scamLogs';
 import detectBan from './logging/detectBan';
 import slashCommandHandler from './slashCommandHandler';
 import contextInteractionHandler from './contextInteractionHandler';
@@ -13,7 +12,6 @@ import buttonHandler from './buttonHandlers/buttonHandler';
 
 export default [
     ready,
-    scamLogs,
     detectBan,
     slashCommandHandler,
     contextInteractionHandler,