+++ src/lib/config.ts
@@ -27,6 +27,7 @@ interface Config {
         FAQ: string;
         FLIGHT_SCHOOL: string;
         KNOWN_ISSUES: string;
+        HONEYPOT: string;
         MOD_ALERTS: string;
         MOD_LOGS: string;
         ROLES: string;
@@ -45,9 +46,15 @@ interface Config {
         [x: string]: string;
     };
     guildId: string;
+    honeypot: {
+        timeoutDurationSeconds: number;
+        deleteWindowSeconds: number;
+    };
     modLogExclude: string[];
     roleAssignmentIds: roleAssignmentId[];
     roleGroups: {
+        STAFF: string[];
+        SUPPORT: string[];
         [x: string]: string[];
     };
     prefixCommandPrefix: string;
@@ -98,7 +105,11 @@ try {
             newRoleGroups[group] = groupRoleIds;
         }
     }
-    parsedConfig.roleGroups = newRoleGroups;
+    const { STAFF, SUPPORT } = newRoleGroups;
+    if (!STAFF || !SUPPORT) {
+        throw new Error('The STAFF and SUPPORT role groups are mandatory');
+    }
+    parsedConfig.roleGroups = { ...newRoleGroups, STAFF, SUPPORT };
     // Parsing Role assignment IDs
     const newRoleAssignmentIds: roleAssignmentId[] = [];
     for (let i = 0; i < roleAssignmentIds.length; i++) {